Cleanup after package installation
It is good to remove the temporary files created by the apt-get package
manager. This makes the image just a little bit smaller.

It is good practice to run all commands in the same invocation of the
"RUN" command, where possible. It reduces the number of layers in the
resulting Docker image.