Automatically use S3 Infrequent Access storage class for older data
Automatically use S3 Infrequent Access storage class for older data
which would have already been in that storage class if it were uploaded
near the date it was taken.

This is useful for the current task of uploading older science data from
the archive.lco.gtn machine as the team moves towards retiring it in
favor of using the AWS S3 Science Archive as our primary data store.