Do not retry HTTP 4xx errors from Archive API
HTTP 4xx errors are classified as client errors, meaning that the client
sent an incorrectly formatted request to the server. These types of
errors are always fatal: retrying would not help, because the problem is
with the request data, not the server.

Therefore, we should never retry after these sorts of errors, they
should always be fatal.