Add valid FITS check for incoming file. Add tests.
Before we create a fits dict from the header or metadata, check if the file is a valid FITS file and handle accordingly. If it is not a valid FITS file, then expect that the user has provided metadata and alert them if they have not. Validate metadata provided against required_headers setting.